Added support for new ePay gateway

+++ lib/epay-payment-soap.php
@@ -12,12 +12,16 @@
  * @copyright ePay Payment Solutions (https://epay.dk) (http://www.epay.dk)
  * @license   ePay Payment Solutions
  */
+// class Epay_Payment_Soap extends epay_payment_api {
 class Epay_Payment_Soap {
 
-	private $pwd = '';
+    private $pwd = '';
+    private $apikey = false;
+    private $posid = false;
 	private $client = null;
 	private $isSubscription = false;
 	private $proxy;
+    private $epay_payment_api;
 
 	/**
 	 * Constructor
@@ -25,8 +29,10 @@ class Epay_Payment_Soap {
 	 * @param mixed $pwd
 	 * @param bool $subscription
 	 */
-	public function __construct( $pwd = '', $subscription = false ) {
+	public function __construct( $pwd = '', $subscription = false,  $apikey = false, $posid = false) {
 		$this->pwd            = $pwd;
+        $this->apikey         = $apikey;
+        $this->posid          = $posid;
 		$this->isSubscription = $subscription;
 		$this->proxy          = new WP_HTTP_Proxy();
 		$options              = array();
@@ -48,6 +54,8 @@ class Epay_Payment_Soap {
 			$service_url,
 			$options
 		);
+
+        $this->epay_payment_api = new epay_payment_api($this->apikey, $this->posid);
 	}
 
 	/**
@@ -66,26 +74,55 @@ class Epay_Payment_Soap {
 	 * @throws Exception
 	 */
 	public function authorize( $merchantnumber, $subscriptionid, $orderid, $amount, $currency, $instantcapture, $group, $email ) {
-		try {
-			$epay_params                   = array();
-			$epay_params['merchantnumber'] = $merchantnumber;
-			$epay_params['subscriptionid'] = $subscriptionid;
-			$epay_params['orderid']        = $orderid;
-			$epay_params['amount']         = (string) $amount;
-			$epay_params['currency']       = $currency;
-			$epay_params['instantcapture'] = $instantcapture;
-			$epay_params['group']          = $group;
-			$epay_params['email']          = $email;
-			$epay_params['pwd']            = $this->pwd;
-			$epay_params['fraud']          = 0;
-			$epay_params['transactionid']  = 0;
-			$epay_params['pbsresponse']    = '-1';
-			$epay_params['epayresponse']   = '-1';
 
-			$result = $this->client->authorize( $epay_params );
-		} catch ( Exception $ex ) {
-			throw $ex;
-		}
+        if(is_numeric($subscriptionid))
+        {
+            try {
+                $epay_params                   = array();
+                $epay_params['merchantnumber'] = $merchantnumber;
+                $epay_params['subscriptionid'] = $subscriptionid;
+                $epay_params['orderid']        = $orderid;
+                $epay_params['amount']         = (string) $amount;
+                $epay_params['currency']       = $currency;
+                $epay_params['instantcapture'] = $instantcapture;
+                $epay_params['group']          = $group;
+                $epay_params['email']          = $email;
+                $epay_params['pwd']            = $this->pwd;
+                $epay_params['fraud']          = 0;
+                $epay_params['transactionid']  = 0;
+                $epay_params['pbsresponse']    = '-1';
+                $epay_params['epayresponse']   = '-1';
+
+                $result = $this->client->authorize( $epay_params );
+            } catch ( Exception $ex ) {
+                throw $ex;
+            }
+        }
+        else
+        {
+            $currency = Epay_Payment_Helper::get_iso_code( $currency, false );
+            $instantcapture = ($instantcapture ? "NO_VOID" : "OFF");
+
+            $result_json = $this->epay_payment_api->authorize($subscriptionid, $amount, $currency, (string) $orderid, $instantcapture, $orderid, Epay_Payment_Helper::get_epay_payment_callback_url( $orderid ));
+            $result_arr = json_decode($result_json, true);
+
+            if(is_array($result_arr))
+            {
+                $result = new stdClass;
+
+                if($result_arr['transaction']['state'] == "PENDING")
+                {
+                    $result->authorizeResult = true;
+                    $result->pbsResponse = 0;
+                }
+                else
+                {
+                    $result->authorizeResult = false;
+                    $result->pbsResponse = -1004;
+                    $result->epayresponse = -1;
+                }
+            }
+        }
 
 		return $result;
 	}
@@ -100,17 +137,37 @@ class Epay_Payment_Soap {
 	 * @throws Exception
 	 */
 	public function delete_subscription( $merchantnumber, $subscriptionid ) {
-		try {
-			$epay_params                   = array();
-			$epay_params['merchantnumber'] = $merchantnumber;
-			$epay_params['subscriptionid'] = $subscriptionid;
-			$epay_params['pwd']            = $this->pwd;
-			$epay_params['epayresponse']   = '-1';
 
-			$result = $this->client->deletesubscription( $epay_params );
-		} catch ( Exception $ex ) {
-			throw $ex;
-		}
+        if(is_numeric($transactionid))
+        {
+            try {
+                $epay_params                   = array();
+                $epay_params['merchantnumber'] = $merchantnumber;
+                $epay_params['subscriptionid'] = $subscriptionid;
+                $epay_params['pwd']            = $this->pwd;
+                $epay_params['epayresponse']   = '-1';
+
+                $result = $this->client->deletesubscription( $epay_params );
+            } catch ( Exception $ex ) {
+                throw $ex;
+            }
+        }
+        else
+        {
+            $result = $this->epay_payment_api->delete_subscription($subscriptionid);
+
+            $result = new stdClass;
+
+            if($result)
+            {
+                $result->deletesubscriptionResult = true;
+            }
+            else
+            {
+                $result->deletesubscriptionResult = false;
+                $result->epayresponse = -1;
+            }
+        }
 
 		return $result;
 	}
@@ -126,19 +183,45 @@ class Epay_Payment_Soap {
 	 * @throws Exception
 	 */
 	public function capture( $merchantnumber, $transactionid, $amount ) {
-		try {
-			$epay_params                   = array();
-			$epay_params['merchantnumber'] = $merchantnumber;
-			$epay_params['transactionid']  = $transactionid;
-			$epay_params['amount']         = (string) $amount;
-			$epay_params['pwd']            = $this->pwd;
-			$epay_params['pbsResponse']    = '-1';
-			$epay_params['epayresponse']   = '-1';
 
-			$result = $this->client->capture( $epay_params );
-		} catch ( Exception $ex ) {
-			throw $ex;
-		}
+        if(is_numeric($transactionid))
+        {
+            try {
+                $epay_params                   = array();
+                $epay_params['merchantnumber'] = $merchantnumber;
+                $epay_params['transactionid']  = $transactionid;
+                $epay_params['amount']         = (string) $amount;
+                $epay_params['pwd']            = $this->pwd;
+                $epay_params['pbsResponse']    = '-1';
+                $epay_params['epayresponse']   = '-1';
+
+                $result = $this->client->capture( $epay_params );
+            } catch ( Exception $ex ) {
+                throw $ex;
+            }
+        }
+        else
+        {
+            $result_json = $this->epay_payment_api->capture($transactionid, $amount);
+            $result_arr = json_decode($result_json, true);
+
+            if(is_array($result_arr))
+            {
+                $result = new stdClass;
+
+                if($result_arr['success'] == true)
+                {
+                    $result->captureResult = true;
+                    $result->pbsResponse = 0;
+                }
+                else
+                {
+                    $result->captureResult = false;
+                    $result->pbsResponse = -1004;
+                    $result->epayresponse = -1;
+                }
+            }
+        }
 
 		return $result;
 	}
@@ -154,19 +237,45 @@ class Epay_Payment_Soap {
 	 * @throws Exception
 	 */
 	public function refund( $merchantnumber, $transactionid, $amount ) {
-		try {
-			$epay_params                   = array();
-			$epay_params['merchantnumber'] = $merchantnumber;
-			$epay_params['transactionid']  = $transactionid;
-			$epay_params['amount']         = (string) $amount;
-			$epay_params['pwd']            = $this->pwd;
-			$epay_params['epayresponse']   = '-1';
-			$epay_params['pbsresponse']    = '-1';
 
-			$result = $this->client->credit( $epay_params );
-		} catch ( Exception $ex ) {
-			throw $ex;
-		}
+        if(is_numeric($transactionid))
+        {
+            try {
+                $epay_params                   = array();
+                $epay_params['merchantnumber'] = $merchantnumber;
+                $epay_params['transactionid']  = $transactionid;
+                $epay_params['amount']         = (string) $amount;
+                $epay_params['pwd']            = $this->pwd;
+                $epay_params['epayresponse']   = '-1';
+                $epay_params['pbsresponse']    = '-1';
+
+                $result = $this->client->credit( $epay_params );
+            } catch ( Exception $ex ) {
+                throw $ex;
+            }
+        }
+        else
+        {
+            $result_json = $this->epay_payment_api->refund($transactionid, $amount);
+            $result_arr = json_decode($result_json, true);
+
+            if(is_array($result_arr))
+            {
+                $result = new stdClass;
+
+                if($result_arr['success'] == true)
+                {
+                    $result->creditResult = true;
+                    $result->pbsResponse = 0;
+                }
+                else
+                {
+                    $result->captureResult = false;
+                    $result->pbsResponse = -1004;
+                    $result->epayresponse = -1;
+                }
+            }
+        }
 
 		return $result;
 	}
@@ -181,17 +290,41 @@ class Epay_Payment_Soap {
 	 * @throws Exception
 	 */
 	public function delete( $merchantnumber, $transactionid ) {
-		try {
-			$epay_params                   = array();
-			$epay_params['merchantnumber'] = $merchantnumber;
-			$epay_params['transactionid']  = $transactionid;
-			$epay_params['pwd']            = $this->pwd;
-			$epay_params['epayresponse']   = '-1';
 
-			$result = $this->client->delete( $epay_params );
-		} catch ( Exception $ex ) {
-			throw $ex;
-		}
+        if(is_numeric($transactionid))
+        {
+            try {
+                $epay_params                   = array();
+                $epay_params['merchantnumber'] = $merchantnumber;
+                $epay_params['transactionid']  = $transactionid;
+                $epay_params['pwd']            = $this->pwd;
+                $epay_params['epayresponse']   = '-1';
+
+                $result = $this->client->delete( $epay_params );
+            } catch ( Exception $ex ) {
+                throw $ex;
+            }
+        }
+        else
+        {
+            $result_json = $this->epay_payment_api->void($transactionid);
+            $result_arr = json_decode($result_json, true);
+
+            if(is_array($result_arr))
+            {
+                $result = new stdClass;
+
+                if($result_arr['success'] == true)
+                {
+                    $result->deleteResult = true;
+                }
+                else
+                {
+                    $result->deleteResult = false;
+                    $result->epayresponse = -1;
+                }
+            }
+        }
 
 		return $result;
 	}
@@ -206,18 +339,39 @@ class Epay_Payment_Soap {
 	 * @throws Exception
 	 */
 	public function get_transaction( $merchantnumber, $transactionid ) {
-		try {
-			$epay_params                   = array();
-			$epay_params['merchantnumber'] = $merchantnumber;
-			$epay_params['transactionid']  = $transactionid;
-			$epay_params['pwd']            = $this->pwd;
-			$epay_params['epayresponse']   = '-1';
 
-			$result = $this->client->gettransaction( $epay_params );
-		} catch ( Exception $ex ) {
-			throw $ex;
-		}
+        if(is_numeric($transactionid))
+        {
+            try {
+                $epay_params                   = array();
+                $epay_params['merchantnumber'] = $merchantnumber;
+                $epay_params['transactionid']  = $transactionid;
+                $epay_params['pwd']            = $this->pwd;
+                $epay_params['epayresponse']   = '-1';
 
+                $result = $this->client->gettransaction( $epay_params );
+            } catch ( Exception $ex ) {
+                throw $ex;
+            }
+        }
+        else        
+        {
+            $result_json = $this->epay_payment_api->payment_info($transactionid);
+            $result_obj = json_decode($result_json);
+            
+            $result = new stdClass;
+
+            if($result_obj->transaction->state == "SUCCESS")
+            {
+                $result->gettransactionResult = true;
+                $result->transactionInformation = $result_obj;
+            }
+            else
+            {
+                $result->gettransactionResult = false;
+                $result->epayresponse = -1;
+            }
+        }
 		return $result;
 	}
